Push on-demand installation into `Host.Provider`
Now that `Host.Provider` accepts at least a `PluginSpec`, it is capable
of encapsulating the logic of attempting a plugin download should an
initial load fail due to a "plugin missing" error. This commit moves
this logic into `Host.Provider` so that it may be removed from several
other parts of the codebase where it is currently duplicated. Moreover,
it (in theory) fixes some bugs/quality of life issues with the parts of
the codebase that do _not_ currently duplicate this code.