Add packageReferences feature flag (#17153)
This plumbs support for the flag into the engine and SDKs and bumps the
minimum version numbers for parameterized SDKs.

Annoyingly it bumps sdkgen to refer to the version about to be released,
but this upsets pip in the test for python so we need to skip that test
for now until we actually release this, then turn it back on.