try to choose a consistent port for debugging
In languages where we need to attach the debugger to a port, we have
some choice which port we can give the user.  For using the debugging
feature on the command line, it is convenient to have the port always
be the same.  Let's try to do that, but go to higher ports if the port
we chose is already in use for some reason.

The ports chosen here are "randomly" chosen high ports that are
unlikely to be used already.