Show the count of unchanged resources (#17257)
The Pulumi CLI does not show details of unchanged resources by default,
since generally it is expected that users only care about what *will*
change when they run their Pulumi program. If the number of unchanged
resources is large, or if no resources have changed, the CLI may spend
long periods of time displaying nothing, or appearing to have hung. The
`--show-sames` argument instructs the CLI to show unchanged resources,
but if the number is large, then this can result in an overly noisy
display, which is not desirable either.

This commit adapts the work of #15975 to display a count of unchanged
resources when `--show-sames` is not enabled. In doing so we can
indicate to the user that progress is being made without introducing
noise or clutter. When the operation is complete, the count is hidden
and replaced with the final summary, which includes the figure as well
as other statistics (such as the number of resources that will change,
and so on). In order to avoid double counting (which appeared to affect
the work in #15975), this commit uses a set keyed by URNs rather than an
`int`.

Closes #15975

Co-authored-by: Pat Gavlin <pat@pulumi.com>