Allow 'pulumi about' in nodejs projects to resolve dependencies from package files by walking up the file tree (#17239)
Resolves https://github.com/pulumi/pulumi/issues/10069: `pulumi about`
fails to find nodejs dependencies when the Pulumi.yaml file is nested
below the directory with the package.json, package-lock.json, and
yarn.lock files, as GetProgramDependencies() in pulumi-language-nodejs
only checks the pulumi program directory (which is the pulumi project
root, in other words the directory with the Pulumi.yaml, unless
overridden) for these files. This commit addresses this bug by adding
logic to GetProgramDependencies to search up the file tree for the
directory with the package files, and uses this directory to compile
dependencies. Also adds unit tests for GetProgramDependencies.

---------

Co-authored-by: Jonathan Davenport <jdavenport@pulumi.com>