Introduce snapshot metadata
This commit introduces a `Metadata` field to state snapshots. The
initial use case for this field will be tracking snapshot integrity
errors, so that better messages can be produced for users and better
diagnostic information preserved for contributors. When this commit is
merged, support for persisting the field will be added to backends
appropriately, before use cases such as the above can be implemented.