Fully generate TypedDict input types by default
https://github.com/pulumi/pulumi/pull/16704 made it so that classes and
TypeDicts are generated for input types by default, but missed some
cases, which lead to referencing some TypedDict types that didn't exist.

This PR fixes that by ensuring that all input types are generated.

A regression test is included that runs pyright in strict mode to ensure
TypedDict types in signatures actually exist.