Fix regression in automation API for deployments
When running deployments via automation API, as the CLI internally
receives log messages from the deployment, it has a heuristic that looks
for the "Pulumi operation" step, and when it sees that step, it makes a
request to Pulumi Cloud to receive the events for the deployment so it
can render those like it does for a local deployment. It does this by
looking for a log line with the header "Pulumi operation".

However, this changed in Pulumi Cloud to be "pulumi <operation>", so the
heuristic no longer works. This means the event stream is not retrieved
and we end up in a situation where the API doesn't work well because
things like the summary info are no longer filled-in at the end of the
operation.

The reason we didn't catch the problem is because the tests are not
currently being run by default.

This change fixes the problem and enables the tests. (We may end up
wanting to run the tests in the a separate CRON workflow.)

A follow-up change will be made to Pulumi Cloud to ensure this doesn't
regress in the future and also consider doing it in a way that is more
stable and doesn't rely on comparing the header of the log line.