Fix _sync_monitor_supports_parameterization (#17300)
Trying to split up https://github.com/pulumi/pulumi/pull/17283 as
something in it keeps failing CI.

This fixes the `_sync_monitor_supports_parameterization` method (which
is currently not being used due to codegen bugs) to refer to the correct
flag.

Also update the test that we need to keep skipping till a later version
with the codegen fixes is released.