allow tests to run in parallel with linting
Currently we make running the tests depend on linting the program.
This occasionally saves some GitHub actions runner time, when the lint
step detects that the build is failing.  However in many cases it's
quite inconvenient, especially if the lint error is really just a lint
error (e.g. because of a long line or something).

In that case the submitter of the PR has to fix the lint first, and
then push those changes, and wait for CI all over again, which can add
a considerable amount of time (especially since CI is kinda slow, so
at least I usually end up doing something else and checking back
later).

In addition it adds to the time it takes to run all tests in total, as
the tests can't be run in parallel with linting (and linting takes
more than a minute).

Allow the tests to run in parallel instead.  This means we can get the
test results early, as long as the code builds fine.  In some
occasions that means some extra time spent on GitHub actions, but
that's time well spent imo.