Add the ability to generate aliased import prefixes to go generator.
Previously, if there were nested arguments in schemas, generation of go
code would simply use the go module name, but when imports are generated
they are potentially aliased to avoid naming conflicts.

This is taken into account at the top level since the aliases are
checked directly, but when traversing downward the argument types are generated
after conversions from the PCL schema data.

This fix adds an optional map to pkgContext that contains the go module
info, which contains the appropriate base path to check against aliases.
We also add aliases to the map that are generated by the import system
(and not just from other go package infos).