Add LanguageInfo to conformance tests (#17315)
This allows a language test to add language specific options to all the
schemas it loads from the conformance test system. This applies to
schemas loaded in GeneratePackage but also GenerateProject.

This is part of fixing https://github.com/pulumi/pulumi/issues/17294
because it turns out that repros iff "inputTypes" is set to
"classes-and-dicts" but during program code generation, which the
current schema overload method we were using for setting toml loading
can't easily support.

This new way covers the toml use case, and can also be used for the
inputTypes scenario as well.