Fix python typed dicts to be enabled by default (#17322)
https://github.com/pulumi/pulumi/pull/16704 correctly defaulted typed
dicts to false for external schemas in sdk-gen but incorrectly turned
them on for external schemas in program-gen iff the schema had any other
python language options.

This fixes the latter to assume that typed dicts can be used in
program-gen. This should be fine because even the old class style could
be passed dict arguments as well, but we'll use the class style if the
schema explicitly says to use it.