Allow passing an initial state at stack creation (#17336)
When creating a stack, there are cases where it would be valuable to be
able to initialise the stack with a known state snapshot (such as one
containing a non-default secrets manager, as discussed in #16890, for
instance). This commit begins adding support by this by extending the
API types for stack creation to take an `UntypedDeployment`. When this
is released, changes to relevant backends (DIY and Pulumi Cloud) can be
implemented to support the new field, at which point the CLI can then be
modified to take advantage of it.