don't run tests using poetry in parallel on windows (#17337)
Poetry uses a cache for packages to speed up installation of the venv.
It seems like the primitives it uses for that are causing issues when
running multiple instances of poetry in parallel that are installing the
same packages. This seems to happen only on Windows. Make these tests
run sequentially on Windows to avoid this test flakyness.

There's an upstream issue for poetry that describes this as well
(https://github.com/python-poetry/poetry/issues/7370)

https://github.com/pulumi/pulumi/issues/17183