Retry token errors immediately (#17338)
If we hit an error trying to refresh the token source (and it's not an
expired token error) then retry right away. This stops two error cases.
Firstly in GetToken where we may return the state token, and secondly on
the timer where the duration between now and the next refresh tick might
push us over the time limit before the token expires.