Add a `property.Path` mechanism, similar to `resource.PropertyPath`
I have been using `property.Value` in providers, but the lack of path support requires
either manual access or using `resource` for access and then `property` for manipulation.

The initial implementation does not support globs. When glob support is added for parity,
it will have it's own set of types to avoid confusion.