Let `Backend.CreateStack` accept an initial state
There are a number of cases where it would be useful to be able to seed
a newly-created stack with an initial state. #17037 and its underlying
issue, for instance, could be resolved by persisting a non-default
secrets manager as part of stack initialisation rather than as a
separate subsequent operation. This commit starts the journey to
enabling this, adding an `initialState` parameter to the
`Backend.CreateStack` interface method.