Allow `Backend`s to define `DefaultSecretManager`s
There are a number of cases where it would be useful to be able to seed
a newly-created stack with an initial state. #17037 and its underlying
issue, for instance, could be resolved by persisting a non-default
secrets manager as part of stack initialisation rather than as a
separate subsequent operation. In such cases, we want to be able to
create a secrets manager for a stack before it has been created.
Presently, the `Stack` interface offers `DefaultSecretManager` for
asking what the default secret manager for a *stack* should be, but
there is no way to ask a `Backend` what the default manager for *stacks
associated with that backend* should be. This commit adds a
`DefaultSecretManager` method to the `Backend` interface, along with
appropriate implementations for DIY and HTTP backends.