Add input keys from diff during import step to warning and error message (#17312)
Currently, when importing a resource via the ResourceOptions, if there
is an input diff, on `pulumi preview` you will see the input keys that
have are creating the diff in the preview plan and a diagnostic warning
message which does not contain the input keys, and on `pulumi up` the
operation will error out, but the error message does not contain the
input keys with the diff. This commit adds the input keys that have a
diff to both the warning and error messages. Also adds unit tests.
Fixes #15606.

---------

Co-authored-by: Jonathan Davenport <jdavenport@pulumi.com>