allow component providers to return more detailed error messages (#17306)
This is based on work from @lunaris and @EronWright, to allow us to
return better error messages from component providers.

The basic idea here is to allow attaching more error details on the RPC
layer, and turn errors into diagnostic *in the engine*, to avoid the
need to do fix the error up in every SDK and pretty print it.

GRPC allows us to attach error details to the returned error to help us
with that. These details can be error details such as specified in the
`error_details` proto that comes with grpc, but can also take any other
shape. Currently we only support pretty printing select types from that
proto, but this can be extended in the future.

On the implementation side, Go has a pretty nice API to create these
errors, which we can just let users use directly. For Python and NodeJS
the API is not so nice for this, so we need to encapsulate the error
into a special exception, and then turn that into a proper GRPC message,
using the magic `grpc-status-details-bin` metadata field.

For both Python and NodeJS I've only implemented one class for errors so
far, as I'm interested in some feedback on the API design first. I'm
wondering if we should just let users specify the fields as tuples, and
then add them to the `error_details` proto? Very open to ideas here.

Closes: https://github.com/pulumi/pulumi/pull/16132

---------

Co-authored-by: Will Jones <will@sacharissa.co.uk>
Co-authored-by: Eron Wright <eron@pulumi.com>