run in local workspace when no PULUMI_ACCESS_TOKEN is set
When testing locally, it's nice to not have to set
`PULUMI_ACCESS_TOKEN` to run tests.  This is both so the tests run
faster (with race detection not turned on, using the local backend is
a lot faster than using the cloud backend), and also so one doesn't
have to remember to log in to the cloud backend.

In addition, we want to enable external contributors to just run the
regular test suite in the future.  For that secrets will not be
available, so running these tests using the local backend is better
than not running them at all (even though they run a little
bit slower in CI with the local backend).