Fix dependency installation for nodejs projects with a main property
After running the package manager installation, we check that a
node_modules folder exists. This check was done relative to the "Program
Directory", however the node_modules folder is created next to
package.json, which itself is in the "Project Directory". When the
Pulumi.yaml specifies a `main` property, these directories are
different, and the installation fails.

We probably did not run into this earlier, because often for nodejs you
don't use Pulumi's `main`, but use the nodejs `main` in package.json.

Fixes https://github.com/pulumi/pulumi/issues/17292