Fix dependency installation for nodejs projects with a main property (#17391)
After running the package manager installation, we check that a
`node_modules` directory exists. This check was done relative to the
"Program Directory", however the `node_modules` directory is created
next to `package.json`, which itself is often in the "Project
Directory". When `Pulumi.yaml` specifies a `main` property, these two
directories are different, and the check fails.

We probably did not run into this earlier, because for nodejs you often
don't use Pulumi's `main`, but instead use the nodejs `main` in
`package.json`.

Fixes https://github.com/pulumi/pulumi/issues/17292