Invalidate GITHUB_TOKEN on 401 (#17351)
Part one of fixing https://github.com/pulumi/pulumi/issues/17217. This
disables and retries the plugin http request without GITUHB_TOKEN if we
get a 401 error.

This should help with cases where people have tokens that don't have
access to public resources. We'll see the 401 and retry without the
token set.

This doesn't fix the other part of #17217 to retry 403 error without the
token.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>