unify prompts and get rid of the erikgeiser/promptkit dependency
We use this dependency in the `config env` commands only, even though
we already have a way to prompt users.  It's better to have a unified
way of prompting for things, and as a bonus we can reduce the amount
of dependencies.

(I originally found this looking at what's pulling in the
atotto/clipboard dependency, and noticed this)