[go/conformance] Fix l2-invoke-simple and l2-invoke-variants tests (#17396)
### Description

This PR fixes l2-invoke-simple and l2-invoke-variants conformance tests
for Go. I was struggling originally to get plain invokes, which return
`(Result, error)` to be rewritten correctly into temp variables but it
broke a few more test examples and didn't go well.

This time, I chose the path of least resistance: enabling
`pcl.PreferOutputVersionedInvokes` which makes invoke functions in PCL
choose the output variant of the invoke regardless of whether their
inputs have anything of type output and the return type `ResultOutput`
is simpler to work with. We also do this for nodejs and python.

The function body of the output-versioned invokes in Go follow an almost
identical code path to the plain ones so this might be Good Enough™ to
assert correctness of invokes.