improve the usage string for the config env add/rm commands
Make it clear which arguments are required for these commands, and
that add takes multiple environments as argument.

Output before:
```
$ pulumi config env add
Adds environments to the end of a stack's import list. Imported environments are merged in order
per the ESC merge rules. The list of stacks behaves as if it were the import list in an anonymous
environment.

Usage:
  pulumi config env add [flags]

[...]
error: requires at least 1 arg(s), only received 0
```

Output after:
```
$ pulumi config env add
Adds environments to the end of a stack's import list. Imported environments are merged in order
per the ESC merge rules. The list of stacks behaves as if it were the import list in an anonymous
environment.

Usage:
  pulumi config env add <environment-name>... [flags]
[...]
error: requires at least 1 arg(s), only received 0
```

The former left me puzzled what argument I need to add, I think the
second one helps the users a lot more to understand what they need to
provide.