Add StringMapMapMap to Go SDK (#17417)
This is to unblock https://github.com/pulumi/pulumi/issues/17415. We add
`StringMapMapMap` to the Go SDK and a test to check that sdk-gen can use
that type.

Really this ought to be a conformance test, but to unblock this quickly
just added a sanity check in the old codegen tests.