[go/conformance] Fix l2-resource-config conformance test (#17409)
### Description

This PR fixes l2-resource-config conformance test in Go. The problem was
that we weren't emitting the version of the package as part the default
resource options unless we have `respectSchemaVersion: true` in the Go
language options in our test schemas. The fix is to emit the exact
version not only when `respectSchemaVersion: true` but also when the
package supports packing and the version is specified (as the case with
the conformance test providers)