fixup! Add the ability to generate aliased import prefixes to go generator.
This has incomplete/failing tests.  It's most of the way there, but this
differs from the situation in #9664 which happened during applies, and
so it seems that while ApplyT is fixed, this kind of overlap is not.

It is aliased now, but the alias is not used (should be pretty simple to
memo it somewhere and use it when generating).