fixup! Fix imports in generated outputs in go converted code.
This test does not pass at all.  The original problem was when ApplyT
was used and it was most likely due to a binding issue, however this
test doesn't cause ApplyT to be generated and so it is difficult to
test.  I'm trying to have a transient dependency so an ApplyT is used to
generate output but it is not working yet.