Persist initial states in HTTP backend (#17373)
There are a number of cases where it would be useful to be able to seed
a newly-created stack with an initial state. #17037 and its underlying
issue, for instance, could be resolved by persisting a non-default
secrets manager as part of stack initialisation rather than as a
separate subsequent operation. Now that `Backend.CreateStack` accepts an
initial state, we can modify implementations to actually persist the new
parameter. This commit does so for the HTTP backend.