Send a secrets manager as initial state (#17387)
There are a number of cases where it would be useful to be able to seed
a newly-created stack with an initial state. #17037 and its underlying
issue, for instance, could be resolved by persisting a non-default
secrets manager as part of stack initialisation rather than as a
separate subsequent operation. With all the pieces in place, this commit
adapts stack creation so that a secrets manager is built *before* stack
creation and sent as part of the initial state to be persisted by the
backend.

Closes #17037 
Fixes #16890 (in tandem with #17236)