Test pyright and mypy in python conformance tests (#17354)
This adds an option to the python language host to emit the typechecker
option during program gen. We set this to mypy and pyright in the
conformance test to ensure our typing annotations are checked during
conformance testing.

Might be there's some way to fit this into `new` and `convert` as well
so people can set they want type checking.