send a value to the abortchan instead of closing it (#17439)
When we're getting an error, we want to abort the source iterator by
closing the abort channel. However, multiple RegisterResource requests
can already be in progress at that time, since they are bbeing run in
parallel. When two of those result in an error, we try to close the
channel twice, which in turn results in a panic because Go doesn't allow
that.

Instead what we should do here is send a value to the abort channel,
which will then poison the iterator, similar to how closing the channel
would poison it.

Fixes https://github.com/pulumi/pulumi/issues/17438