fix dll loading errors in windows CI
Loading the user32 DLL sometimes fails in CI for unknown reasons.
This seems to come from the atotto/clipboard package, which loads the
DLL globally, when it's being initialized. This commit attempts to
work around that by using a version that lazy loads the DLL, similar
to a previous fix that was done in https://github.com/xanzy/ssh-agent/pull/10,
where we faced a different problem.

This replace should obviously be updated once upstream merges my
change.