Fix regression in automation API for deployments (#17334)
When running deployments via automation API, as the CLI internally
receives log messages from the deployment, it has a heuristic that looks
for the "Pulumi operation" step, and when it sees that step, it makes a
request to Pulumi Cloud to receive the events for the deployment so it
can render those like it does for a local deployment. It does this by
looking for a log line with the header `"Pulumi operation"`.

However, this changed in Pulumi Cloud to be `"pulumi <operation>"`, so
the heuristic no longer works. This means the event stream is not
retrieved and we end up in a situation where the API doesn't work well
because things like the summary is no longer available at the end of the
operation, since a summary event was not received.

The reason we didn't catch the problem is because the tests are not
currently being run by default. Unfortunately, the tests add a
significant amount of time it takes to run the tests. In some cases
tests that used to take ~16/17min in CI now take 55 min and timeout.
https://github.com/pulumi/pulumi/issues/17451 tracks enabling the tests.

A follow-up change will be made to Pulumi Cloud to ensure this doesn't
regress in the future, and an additional change will be considered to do
this in a way that doesn't rely on comparing the header of the log line.

Fixes #17155