Preserve ordering when editing stack config.
This fix does not append new nodes when encountered but instead adds
them to the end after all nodes are processed.

I also took to opportunity to rename new to newNode to not shadow the
keyword.  Suprising that go even allows this.

Also fixes existing tests that did not previously preserve order but did
preserve comments.

Fixes: #14860