fix dll loading errors in windows CI (#17410)
Loading the user32 DLL sometimes fails in CI for unknown reasons. This
seems to come from the atotto/clipboard package, which loads the DLL
globally, when it's being initialized. This commit attempts to work
around that by using a version that simply doesn't load the DLL at all,
to see if that fixes the issues.

We could potentially do a bit better, and actually hack the library so
it only tries to load the DLL when the clipboard is actually being used,
but I want to see if this works first.

Fixes https://github.com/pulumi/pulumi/issues/14240