Return pruned resources from `Snapshot.Prune`
The `Snapshot.Prune` method was introduced in #17408 in order to remove
dangling references from faulty state snapshots. In this commit we
extend its capabilities, returning a list of the results that were
pruned and/or rewritten. In doing so we enable callers (such as the
futuer `state repair` command) to print out a summary of the changes
affected by pruning if they wish.