Extend `promptUser` to accept a list of `survey` options
The `survey` library that we use for prompting in the Pulumi CLI
supports several options, such as being able to specify the streams that
should be used for standard input, standard output, and standard error.
We'd like to be able to take advantage of these options in order to
ensure that our commands our testable. This commit extends `promptUser`,
a wrapper function we use, to accept and forward such options so that we
can do just that.